+++ src/API/FeedIssues.php
@@ -56,7 +56,7 @@ class FeedIssues extends VendorAPI {
 	public function get_feed_issues( WP_REST_Request $request ) {
 
 		try {
-			$feed_id = Pinterest\ProductSync::get_registered_feed_id();
+			$feed_id = Pinterest\FeedRegistration::get_registered_feed_id();
 			if ( ! Pinterest\ProductSync::is_product_sync_enabled() || ! $feed_id ) {
 				return array( 'lines' => array() );
 			}
